+++ src/networks/types/AVAXTEST.js
@@ -1,0 +1,15 @@
+import avax from '@/assets/images/networks/avax-c.svg';
+
+export default {
+  name: 'AVAXTEST',
+  name_long: 'Avalanche C-Chain Testnet',
+  homePage: 'https://www.avalabs.org/',
+  blockExplorerTX: 'https://cchain.explorer.avax-test.network/tx/[[txHash]]',
+  blockExplorerAddr:
+    'https://cchain.explorer.avax-test.network/address/[[address]]',
+  chainID: 43113,
+  tokens: [],
+  contracts: [],
+  icon: avax,
+  currencyName: 'C-AVAX'
+};