+++ src/layouts/InterfaceLayout/containers/SwapContainer/SwapContainer.vue
@@ -1177,6 +1177,7 @@ export default {
 
         return enoughToContinue;
       }
+      return true;
     },
     intermediateGasCheck() {
       if (this.fromCurrency === 'ETH' && this.lastFeeEstimate.gt(0)) {