+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Bug
+
+* Fix checking gas when swap from is not on ethereum [#2879](https://github.com/MyEtherWallet/MyEtherWallet/pull/2879)
+
 ### Devop
 
 * consolidate node PRs [#2883](https://github.com/MyEtherWallet/MyEtherWallet/pull/2883)